Refactor host pool related functions
