+++ include/objects.h
@@ -40,9 +40,6 @@ extern PgPreparedStatement *prepared_statements;
 
 extern unsigned long long int last_pgsocket_id;
 
-PgHost *pg_create_host(const char *host_name, int port);
-bool parse_database_hosts(PgDatabase *db);
-
 PgDatabase *find_peer(int peer_id);
 PgDatabase *find_database(const char *name);
 PgDatabase *find_or_register_database(PgSocket *connection, const char *name);
